Add useListController hook
