+++ packages/ra-core/src/controller/index.ts
@@ -1,7 +1,7 @@
 import {
     getListControllerProps,
     sanitizeListRestProps,
-} from './ListController';
+} from './useListController';
 import CreateController from './CreateController';
 import EditController from './EditController';
 import ListController from './ListController';
@@ -10,6 +10,8 @@ import useRecordSelection from './useRecordSelection';
 import useVersion from './useVersion';
 import useSortState from './useSortState';
 import usePaginationState from './usePaginationState';
+import useListController from './useListController';
+import { useCheckMinimumRequiredProps } from './checkMinimumRequiredProps';
 export {
     getListControllerProps,
     sanitizeListRestProps,
@@ -17,6 +19,8 @@ export {
     EditController,
     ListController,
     ShowController,
+    useCheckMinimumRequiredProps,
+    useListController,
     useRecordSelection,
     useVersion,
     useSortState,
