+++ packages/ra-core/src/controller/useListController.tsx
@@ -1,0 +1,215 @@
+import { isValidElement, ReactElement, useMemo } from 'react';
+import inflection from 'inflection';
+
+import { SORT_ASC } from '../reducer/admin/resource/list/queryReducer';
+import { ListParams } from '../actions/listActions';
+import { useCheckMinimumRequiredProps } from './checkMinimumRequiredProps';
+import { Sort } from '../types';
+import { Location } from 'history';
+import { useTranslate } from '../i18n';
+import useListParams from './useListParams';
+import useGetList from './../fetch/useGetList';
+import useRecordSelection from './useRecordSelection';
+import useVersion from './useVersion';
+
+export interface Props {
+    // the props you can change
+    filter?: object;
+    filters?: ReactElement<any>;
+    filterDefaultValues?: object;
+    pagination?: ReactElement<any>;
+    perPage?: number;
+    sort?: Sort;
+    // the props managed by react-admin
+    basePath: string;
+    debounce?: number;
+    hasCreate?: boolean;
+    hasEdit?: boolean;
+    hasList?: boolean;
+    hasShow?: boolean;
+    location: Location;
+    path?: string;
+    query: ListParams;
+    resource: string;
+    [key: string]: any;
+}
+
+const defaultSort = {
+    field: 'id',
+    order: SORT_ASC,
+};
+
+/**
+ * Prepare data for the List view
+ *
+ * @param {Object} props The props passed to the Lsit component.
+ *
+ * Props should contain:
+ *   - title
+ *   - perPage
+ *   - sort
+ *   - filter (the permanent filter to apply to the query)
+ *   - actions
+ *   - filters (a React component used to display the filter form)
+ *   - pagination
+ *
+ * @return {Object} controllerProps Fetched and computed data for the List view
+ *
+ * controllerProps contain:
+ *
+ */
+const useListController = (props: Props) => {
+    useCheckMinimumRequiredProps(
+        'List',
+        ['basePath', 'location', 'resource'],
+        props
+    );
+
+    const {
+        basePath,
+        resource,
+        hasCreate,
+        location,
+        filterDefaultValues,
+        sort = defaultSort,
+        perPage = 10,
+        filter,
+        debounce = 500,
+    } = props;
+
+    if (filter && isValidElement(filter)) {
+        throw new Error(
+            '<List> received a React element as `filter` props. If you intended to set the list filter elements, use the `filters` (with an s) prop instead. The `filter` prop is internal and should not be set by the developer.'
+        );
+    }
+    const translate = useTranslate();
+    const version = useVersion();
+
+    const [query, queryModifiers] = useListParams({
+        resource,
+        location,
+        filterDefaultValues,
+        sort,
+        perPage,
+        debounce,
+    });
+
+    const [selectedIds, selectionModifiers] = useRecordSelection(resource);
+
+    const { data, ids, total, loading, loaded } = useGetList(
+        resource,
+        {
+            page: query.page,
+            perPage: query.perPage,
+        },
+        { field: query.sort, order: query.order },
+        { ...query.filter, ...filter },
+        {
+            version,
+            onFailure: {
+                notification: {
+                    body: 'ra.notification.http_error',
+                    level: 'warning',
+                },
+            },
+        }
+    );
+
+    if (!query.page && !(ids || []).length && query.page > 1 && total > 0) {
+        // query for a page that doesn't exist, check the previous page
+        queryModifiers.setPage(query.page - 1);
+    }
+
+    const currentSort = useMemo(
+        () => ({
+            field: query.sort,
+            order: query.order,
+        }),
+        [query.sort, query.order]
+    );
+
+    const resourceName = translate(`resources.${resource}.name`, {
+        smart_count: 2,
+        _: inflection.humanize(inflection.pluralize(resource)),
+    });
+    const defaultTitle = translate('ra.page.list', {
+        name: resourceName,
+    });
+
+    return {
+        basePath,
+        currentSort,
+        data,
+        defaultTitle,
+        displayedFilters: query.displayedFilters,
+        filterValues: query.filterValues,
+        hasCreate,
+        ids,
+        isLoading: loading,
+        loadedOnce: loaded,
+        onSelect: selectionModifiers.select,
+        onToggleItem: selectionModifiers.toggle,
+        onUnselectItems: selectionModifiers.clearSelection,
+        page: query.page,
+        perPage: query.perPage,
+        resource,
+        selectedIds,
+        setFilters: queryModifiers.setFilters,
+        hideFilter: queryModifiers.hideFilter,
+        showFilter: queryModifiers.showFilter,
+        setPage: queryModifiers.setPage,
+        setPerPage: queryModifiers.setPerPage,
+        setSort: queryModifiers.setSort,
+        total,
+        version,
+    };
+};
+
+export const injectedProps = [
+    'basePath',
+    'currentSort',
+    'data',
+    'defaultTitle',
+    'displayedFilters',
+    'filterValues',
+    'hasCreate',
+    'hideFilter',
+    'ids',
+    'isLoading',
+    'loadedOnce',
+    'onSelect',
+    'onToggleItem',
+    'onUnselectItems',
+    'page',
+    'perPage',
+    'refresh',
+    'resource',
+    'selectedIds',
+    'setFilters',
+    'setPage',
+    'setPerPage',
+    'setSort',
+    'showFilter',
+    'total',
+    'version',
+];
+
+/**
+ * Select the props injected by the useListController hook
+ * to be passed to the List children need
+ * This is an implementation of pick()
+ */
+export const getListControllerProps = props =>
+    injectedProps.reduce((acc, key) => ({ ...acc, [key]: props[key] }), {});
+
+/**
+ * Select the props not injected by the useListController hook
+ * to be used inside the List children to sanitize props injected by List
+ * This is an implementation of omit()
+ */
+export const sanitizeListRestProps = props =>
+    Object.keys(props)
+        .filter(propName => !injectedProps.includes(propName))
+        .reduce((acc, key) => ({ ...acc, [key]: props[key] }), {});
+
+export default useListController;
